Changes made: - Restructuring to classes & tools - Removed unnessary code in Data Analyser - Restructured cleaner to contain the correct functions and handle the whole dataframe - Implemented first version of example script for end user in pipeline - Converted wordcloudgen, fex & geotagger into class structure and testing in pipeline - WIP: testing of new class structure in pipeline.py - TODO: sentiment analyser & user classification still to convert into class structure
