+++ tools/analyser.py
@@ -1,0 +1,4 @@
+class sentimentAnalyser:
+
+    def __init__(self, df):
+        self.df = df
