+++ tools/basic_headers.csv
@@ -1,0 +1,1 @@
+created_at,id,text,entities,source,user.id,user.screen_name,user.location,user.followers_count,user.friends_count,user.verified,user.statuses_count,geo,coordinates,place,user_location2,retweet_count,tweet_favorite_count,tweet_favoritedtweet_retweeted
