+++ tools/classifier.py
@@ -1,0 +1,4 @@
+class userclassifier:
+
+    def __init__(self, df):
+        self.df = df
